Reloading, Pause Menu, Drag and Drop
Guns have set ammo, fire rate, and need reloading
They reload automatically. Probably more fun if that's the case.
Remembering to press r is a skill, but it's more relaxing if you don't
have to.

Can press space to pause the game. Will be triggered by something else
later.
Can move dice into slots to give buffs or whatever.
1 die per slot.
Slots empty when you move the die
Slots know which dice are in them
This is most of what I need for the dice portion of the game.