Roll Dice, Dice Reset
Extended the drag and drop
Now the items are dice that can be rolled and have a value
The item slots can read the value of the die placed in them.
Dice and slots get reset on a new pause.

Prefabs
Can almost spawn in dice and slots as the player gains more through
gameplay.

DiceMenuController
Handles telling stuff when to show, when dice should roll, when the
results should be evaluated.

Player
Player could shoot a bullet while paused. Now they can't.

Bugs
Reference to UIDie in ItemSlot. ItemSlot should be agnostic of the
inheritance of DragDrop, but it isn't. I might fix it. It doesn't really
matter for a game jam I don't think. I doubt there will be DragDrop items
that aren't UIDie.