Dice Pickups
Enemies drop dice sometimes. Pick it up to have an extra die for the next
pause. After that, these pickup dice are deleted whether you use them or
not.
Bosses have the same drop chance as basic enemies.

Dice Pool
Made it neater. Dice are shown in a little grid that expands as you get
more dice.
Doing this broke the dice fitting in to the slot, but I fixed it. Layout
groups and anchor nonsense.