Grid Buff, Grenade Launcher
Put in 2 dice to light up a grid square. Get a buff while standing inside
of it.

Grenade Launcher
Launch a grenade that explodes in a small area.

UI
Put an X and Y in the combo slots so it's clearer.

Bug Fixes
When you get bumped, you can drift for a while sometimes. I might have
fixed it by setting velocity to 0. Who know? I always have this problem with physics.